apps summary basic setup
