+++ src/client/App.tsx
@@ -2,7 +2,9 @@ import React from 'react';
 import { Routes, Route } from 'react-router-dom';
 import SignInForm from './components/SignInForm';
 import SignUpForm from './components/SignUpForm';
-// import Applications from './components/Applications';
+import Applications from './components/Applications';
+// import AddApplication from './components/AddApplication';
+// import ApplicationDetails from './components/ApplicationDetails';
 import './styles/main.scss';
 
 
@@ -13,8 +15,9 @@ const App = () => {
       <Routes>
         <Route path='/' element={<SignInForm />} />
         <Route path='/signup' element={<SignUpForm />} />
-        {/* <Route path='/applications' Component={Applications} /> */}
-        {/* <Route path='/appDetails' Component={AppDetails} /> */}
+        <Route path='/applications' element={<Applications />} />
+        {/* <Route path='/applications/new' element={<AddApplication />} /> */}
+        {/* <Route path='/applications/:id' element={<ApplicationDetails />} /> */}
         {/* <Route path='/interviews' Component={Interviews} />
         <Route path='/questions' Component={Questions} /> */ }
       </Routes>
