+++ src/client/components/ApplicationDetails.tsx
@@ -2,18 +2,12 @@ import React, { useContext, useState, useEffect } from 'react';
 import { useNavigate, Link, Route } from 'react-router-dom';
 import App from '../App';
 
-export interface Application {
-  id: number;
-  position: string;
-  appDate: string;
-  mostRecentContact: string;
-  company: string;
-  status: string;
-  details: string;
+const ApplicationDetails = () => {
+  return (
+    <div>
+      <p>This is application details page</p>
+    </div>
+  )
 }
 
-export interface Status {
-  total: number;
-  notStarted: number;
-  inProgress: number;
-}
+export default ApplicationDetails;
