+++ src/client/styles/main.scss
@@ -18,8 +18,8 @@
   justify-content: space-between;
   align-items: center;
   padding: 0 20px;
-  background-color: $main-color;
-  color: rgb(222, 117, 25);
+  background-color: mediumspringgreen;
+  color: rgb(222, 25, 183);
   height: 60px;
 }
 
@@ -29,12 +29,12 @@
   justify-content: space-between;
   align-items: center;
   padding: 0 20px;
-  background-color: $main-color;
+  background-color: violet;
   color: rgb(27, 162, 85);
   height: 60px;
 }
 .headerCol {
   display: flex;
   align-items: center;
-  color: orange
+  color: rgb(1, 4, 2)
 }
