+++ webpack.config.js
@@ -46,7 +46,7 @@ module.exports = {
     },
     port: 8080,
     proxy: {
-      '/api': {
+      '/api/**': {
         target: 'http://localhost:3000',
         pathRewrite: {'^/api' : ''}
       }
